cli: extract doctor cli support and tests
